move calculations and conversion to own module
