+++ nobubo/utils.py
@@ -18,10 +18,7 @@
 """
 Various helper classes and methods.
 """
-import math
 from dataclasses import dataclass
-
-import PyPDF2
 
 
 @dataclass
@@ -68,44 +65,3 @@ class Point:
     y: float
 
 
-def calculate_pages_needed(layout: Layout, n_up_factor: Factor) -> int:
-    return math.ceil(layout.columns/n_up_factor.x) * math.ceil(layout.rows/n_up_factor.y)
-
-
-def calculate_page_dimensions(page: PyPDF2.pdf.PageObject) -> (float, float):
-    """
-    Calculates the x, y value for the offset in default user space units as defined in the pdf standard.
-    Uses the cropBox value, since this is the area visible to the printer.
-    :param page: A pattern page.
-    :return: list with x, y value.
-    """
-    return round(float(page.cropBox[2])-float(page.cropBox[0]), 2), round(float(page.cropBox[3])-float(page.cropBox[1]), 2)
-
-
-def convert_to_userspaceunits(width_height: [int, int]) -> PageSize:
-    """
-    Converts a page's physical width and height from millimeters to default user space unit,
-    which are defined in the pdf standard as 1/72 inch.
-
-    :param width_height: Width and height of the physical page in millimeters (mm),
-    on which the pattern will be printed.
-    :return: Width and height of the physical page in default user space units.
-    """
-    # 1 mm = 5/127 inches = 0.03937 inches;  1/72 inch = 0.013888889
-    # conversion factor = 5/127 / 1/72 = 360/127 = 2.834645669
-    conversion_factor = 2.834645669
-
-    return PageSize(width=(round(width_height[0] * conversion_factor, 3)),
-                    height=(round(width_height[1] * conversion_factor, 3)))
-
-
-def calculate_nup_factors(output_layout: [int], input_properties: PDFProperties) -> Factor:
-    output_papersize = convert_to_userspaceunits(output_layout)
-    x_factor = int(output_papersize.width // input_properties.pagesize.width)
-    y_factor = int(output_papersize.height // input_properties.pagesize.height)
-    return Factor(x=x_factor, y=y_factor)
-
-
-def convert_to_mm(output_layout: str) -> [int, int]:
-    ol_in_mm = output_layout.split("x")
-    return [int(x) for x in ol_in_mm]
